Refactor telemetry file structure and imports
