--- src/telemetry/index.ts
@@ -1,2 +1,0 @@
-export { DefaultTelemetry } from './default-telemetry.js';
-export type { Telemetry } from './types.js';
